# docs/swagger/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@openauth.local
    name: API Support
    url: https://github.com/hanyouqing/OpenAuth/issues
  description: OpenAuth - Open-source Identity and Access Management (IAM) platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/hanyouqing/OpenAuth
  title: OpenAuth API
  version: "1.0"
paths:
...
  /oauth2/authorize:
    get:
      description: OAuth 2.0 Authorization Code Flow authorization endpoint
      parameters:
      - description: Response type (code)
        in: query
        name: response_type
        required: true
        type: string
      - description: Client ID
        in: query
        name: client_id
        required: true
        type: string
      - description: Redirect URI
        in: query
        name: redirect_uri
        required: true
        type: string
      - description: Requested scopes
        in: query
        name: scope
        type: string
      - description: State parameter
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect to authorization page or redirect_uri
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: OAuth 2.0 Authorization
      tags:
      - sso
  /oauth2/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: OAuth 2.0 Resource Owner Password Credentials Flow
      parameters:
      - description: Grant type (password)
        in: formData
        name: grant_type
        required: true
        type: string
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Client ID (optional)
        in: formData
        name: client_id
        type: string
      - description: Client secret (optional)
        in: formData
        name: client_secret
        type: string
      - description: Requested scopes
        in: formData
        name: scope
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Access token and refresh token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: OAuth 2.0 Password Credentials
      tags:
      - sso
  /oauth2/userinfo:
    get:
      description: Get user information using access token (OIDC UserInfo endpoint)
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: OIDC UserInfo
      tags:
      - sso
...
  /saml/metadata:
    get:
      description: Get SAML 2.0 Entity Descriptor metadata
      parameters:
      - description: Application ID
        in: query
        name: application_id
        required: true
        type: integer
      produces:
      - application/xml
      responses:
        "200":
          description: SAML Metadata XML
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: SAML 2.0 Metadata
      tags:
      - sso
  /saml/slo:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: SAML 2.0 Single Logout endpoint
      parameters:
      - description: SAML Logout Request
        in: query
        name: SAMLRequest
        type: string
      - description: SAML Logout Response
        in: formData
        name: SAMLResponse
        type: string
      - description: Relay state
        in: query
        name: RelayState
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: SAML Logout Response (POST form or redirect)
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: SAML 2.0 SLO
      tags:
      - sso
  /saml/sso:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: SAML 2.0 Single Sign-On endpoint (supports SP-initiated and IdP-initiated)
      parameters:
      - description: SAML Request (SP-initiated)
        in: query
        name: SAMLRequest
        type: string
      - description: SAML Response
        in: formData
        name: SAMLResponse
        type: string
      - description: Relay state
        in: query
        name: RelayState
        type: string
      produces:
      - text/html
      - application/json
      responses:
        "200":
          description: SAML Response (POST form or redirect)
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: SAML 2.0 SSO
      tags:
      - sso
  /sessions:
    delete:
      description: Delete all sessions for current user
      produces:
      - application/json
      responses:
        "200":
          description: All sessions deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete all sessions
      tags:
      - sessions
    get:
      description: Get paginated list of user sessions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session list
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List sessions
      tags:
      - sessions
  /sessions/{id}:
    delete:
      description: Delete a session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete session
      tags:
      - sessions
  /sessions/active/count:
    get:
      description: Get count of active sessions for current user
      produces:
      - application/json
      responses:
        "200":
          description: Active session count
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get active session count
      tags:
      - sessions
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"